Matlab scripts + windows fixes

+++ python_server/gdal_interfaces.py
@@ -65,8 +65,8 @@ class GDALInterface(object):
             u = xgeo - self.geo_transform_inv[0]
             v = ygeo - self.geo_transform_inv[3]
             # FIXME this int() is probably bad idea, there should be half cell size thing needed
-            xpix = float(self.geo_transform_inv[1] * u + self.geo_transform_inv[2] * v)
-            ylin = float(self.geo_transform_inv[4] * u + self.geo_transform_inv[5] * v)
+            xpix = int(self.geo_transform_inv[1] * u + self.geo_transform_inv[2] * v)
+            ylin = int(self.geo_transform_inv[4] * u + self.geo_transform_inv[5] * v)
             # print("1. ",self.geo_transform_inv[1] * u + self.geo_transform_inv[2] * v)
             # look the value up
             v = self.points_array[ylin, xpix]
